+++ CHANGELOG.md
@@ -2,11 +2,19 @@
 
 All notable changes to `cookies-popup` will be documented in this file
 
+## 1.2.2 - 2024-04-24
+
+- Google consent mode improvement
+- Add if `$firstMeasurementId` to gtag calls
+
+> Add `google_ads_data_redaction` in your `config/cookies-popup.php`
+
 ## 1.2.1 - 2024-04-11
 
 - Google consent mode improvement
 
 > Add `google_url_passthrough`, `google_wait_for_update` and `custom_styles.overlay_z_index` in your `config/cookies-popup.php`
+
 ## 1.2.0 - 2024-04-04
 
 - Google consent mode
@@ -18,7 +26,8 @@ All notable changes to `cookies-popup` will be documented in this file
 - Added add-inline-styles to config
 - New Method getHeadStyles()
 - Removed autocomplete=off on checkboxes
-- 
+-
+
 ```php
 {!! \Itemvirtual\CookiesPopup\CookiesPopup::getHeadStyles() !!}
 ```